VS adjustments, conversions, casts, etc.

+++ g_misc.c
@@ -944,7 +944,7 @@ void barrel_explode (edict_t *self)
 	ThrowDebris (self, "models/objects/debris1/tris.md2", spd, org);
 
 	// bottom corners
-	spd = 1.75 * (float)self->dmg / 200.0;
+	spd = 1.75 * self->dmg / 200;
 	VectorCopy (self->absmin, org);
 	ThrowDebris (self, "models/objects/debris3/tris.md2", spd, org);
 	VectorCopy (self->absmin, org);
@@ -959,7 +959,7 @@ void barrel_explode (edict_t *self)
 	ThrowDebris (self, "models/objects/debris3/tris.md2", spd, org);
 
 	// a bunch of little chunks
-	spd = 2 * self->dmg / 200;
+	spd = (float)(2 * self->dmg / 200);
 	org[0] = self->s.origin[0] + crandom() * self->size[0];
 	org[1] = self->s.origin[1] + crandom() * self->size[1];
 	org[2] = self->s.origin[2] + crandom() * self->size[2];
@@ -1678,7 +1678,7 @@ void target_string_use (edict_t *self, edict_t *other, edict_t *activator)
 	{
 		if (!e->count)
 			continue;
-		n = e->count - 1;
+		n = (size_t)e->count - 1;
 		if (n > len)
 		{
 			e->s.frame = 12;
