+++ p_hud.c
@@ -380,7 +380,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
             "xv 160 yv 40 string2 \"------------------- \" "
             );
 
-        j = strlen(string2);
+        j = (int)strlen(string2);
         if (stringlength + j <= 1024)
         {
             strcpy (string + stringlength, string2);
@@ -419,7 +419,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
 			}
 
             
-            j = strlen(string2);
+            j = (int)strlen(string2);
             if (stringlength + j > 1024)
                 break;
             strcpy (string + stringlength, string2);
@@ -442,7 +442,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
             {
                 Com_sprintf (entry, sizeof(entry),
                     "xv %i yv %i picn %s ",x+32, y, tag);
-                j = strlen(entry);
+                j = (int)strlen(entry);
                 if (stringlength + j > 1024)
                     break;
                 strcpy (string + stringlength, entry);
@@ -456,7 +456,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
                 cl->ping,
                 (level.framenum - cl->resp.enterframe)/600);    
 
-            j = strlen(entry);
+            j = (int)strlen(entry);
             if (stringlength + j > 1024)
                 break;
             strcpy (string + stringlength, entry);
@@ -469,7 +469,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
                     x+32+80, y+24,
                     stats_get(cl_ent, STATS_CAPTURES));
 
-                j = strlen(string2);
+                j = (int)strlen(string2);
                 if (stringlength + j > 1024)
                     break;
                 strcpy (string + stringlength, string2);
@@ -481,7 +481,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
                 Com_sprintf (string2, sizeof(string2),
                     "xv %d yv %d picn dmvpicon ",
                     x, y);
-                j = strlen(string2);
+                j = (int)strlen(string2);
                 if (stringlength + j > 1024)
                     break;
                 strcpy (string + stringlength, string2);
@@ -492,7 +492,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
                 Com_sprintf (string2, sizeof(string2),
                     "xv %d yv %d picn omvpicon ",
                     x, y);
-                j = strlen(string2);
+                j = (int)strlen(string2);
                 if (stringlength + j > 1024)
                     break;
                 strcpy (string + stringlength, string2);
@@ -537,7 +537,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
 					cl->ping > 999 ? 999 : cl->ping);
 			}
 
-            j = strlen(string2);
+            j = (int)strlen(string2);
             if (stringlength + j > 1024)
                 break;
             strcpy (string + stringlength, string2);
@@ -561,7 +561,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
             {
                 Com_sprintf (entry, sizeof(entry),
                     "xv %i yv %i picn %s ",x+32, y, tag);
-                j = strlen(entry);
+                j = (int)strlen(entry);
                 if (stringlength + j > 1024)
                     break;
                 strcpy (string + stringlength, entry);
@@ -574,7 +574,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
                 x, y, bluesorted[i], stats_get(cl_ent, STATS_SCORE),
                 cl->ping, (level.framenum - cl->resp.enterframe)/600);  
 
-            j = strlen(entry);
+            j = (int)strlen(entry);
             if (stringlength + j > 1024)
                 break;
             strcpy (string + stringlength, entry);
@@ -587,7 +587,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
                     x+32+80, y+24,
                     stats_get(cl_ent, STATS_CAPTURES));
 
-                j = strlen(string2);
+                j = (int)strlen(string2);
                 if (stringlength + j > 1024)
                     break;
                 strcpy (string + stringlength, string2);
@@ -598,7 +598,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
                 Com_sprintf (string2, sizeof(string2),
                     "xv %d yv %d picn dmvpicon ",
                     x, y);
-                j = strlen(string2);
+                j = (int)strlen(string2);
                 if (stringlength + j > 1024)
                     break;
                 strcpy (string + stringlength, string2);
@@ -609,7 +609,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
                 Com_sprintf (string2, sizeof(string2),
                     "xv %d yv %d picn omvpicon ",
                     x, y);
-                j = strlen(string2);
+                j = (int)strlen(string2);
                 if (stringlength + j > 1024)
                     break;
                 strcpy (string + stringlength, string2);
@@ -660,7 +660,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
 			y += 8;
 		}
 
-		j = strlen(string2);
+		j = (int)strlen(string2);
 		if (stringlength + j <= 1024)
 		{
 			strcpy (string + stringlength, string2);
@@ -684,7 +684,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
 				y += 8;
 			}
 
-			j = strlen(string2);
+			j = (int)strlen(string2);
 			if (stringlength + j <= 1024)
 			{
 				strcpy (string + stringlength, string2);
@@ -707,7 +707,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
 				y += 8;
 			}
 
-			j = strlen(string2);
+			j = (int)strlen(string2);
 			if (stringlength + j <= 1024)
 			{
 				strcpy (string + stringlength, string2);
@@ -741,7 +741,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
 						y += 8;
 				}
 
-				j = strlen(string2);
+				j = (int)strlen(string2);
 				if (stringlength + j <= 1024)
 				{
 					strcpy (string + stringlength, string2);
@@ -764,7 +764,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
 					y += 8;
 				}
 
-				j = strlen(string2);
+				j = (int)strlen(string2);
 				if (stringlength + j <= 1024)
 				{
 					strcpy (string + stringlength, string2);
@@ -799,7 +799,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
             x, y, entry
             );
 
-        j = strlen(string2);
+        j = (int)strlen(string2);
         if (stringlength + j <= 1024)
         {
             strcpy (string + stringlength, string2);
@@ -852,7 +852,7 @@ void DeathmatchScoreboardMessage (edict_t *ent, edict_t *killer)
         250, 4
         );
 
-        j = strlen(string2);
+        j = (int)strlen(string2);
         if (stringlength + j < 1024)
         {       
             strcpy (string + stringlength, string2);
@@ -1028,7 +1028,7 @@ void CTFSquadboardMessage (edict_t *ent, edict_t* killer) // ADC
 		{
 			squad = sortedClients[i]->pers.squad;
 
-			sprintf(entry+strlen(entry),
+			sprintf(entry+(int)strlen(entry),
 				"xv 0 yv %d string \"%s\" ",
 				42 + i * 8 + numCategoryLines * 8,
 				sortedClients[i]->pers.squad);
@@ -1049,7 +1049,7 @@ void CTFSquadboardMessage (edict_t *ent, edict_t* killer) // ADC
 		// for a netname. We want the names padded with spaces
 		// to make the status line up.
 
-		sprintf(entry+strlen(entry),
+		sprintf(entry+(int)strlen(entry),
 			"xv 0 yv %d %s \"   %-*s %s\" ",
 			42 + i * 8 + numCategoryLines * 8,
 			ready ? readyString : notReadyString,
